requirements.txt
python-telegram-bot>=20.7,<22
py-clob-client>=0.18.0
aiosqlite>=0.20.0
aiohttp>=3.9.0
web3>=6.15.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
